Refactor pact-jvm-mock-spring: restructure dependencies, remove obsolete test files, and add new test module
- Updated `build.gradle.kts` to include the new `pact-jvm-mock-test` module and adjusted dependencies accordingly.
- Removed several obsolete test files and classes related to user and shopping services to streamline the codebase.
- Cleaned up imports and dependencies in the remaining test files to ensure consistency and maintainability.
- This refactor aims to enhance the organization of the test structure and improve overall project clarity.
