--- pact-jvm-mock-spring/src/test/kotlin/io/github/ludorival/pactjvm/mock/spring/providers/userservice/UserPreferences.kt
@@ -1,4 +1,0 @@
-package io.github.ludorival.pactjvm.mock.spring.providers.userservice
-
-data class UserPreferences(val preferredShoppingListId: Long)
-
