--- pact-jvm-mock-spring/src/test/kotlin/io/github/ludorival/pactjvm/mock/spring/providers/userservice/UserProfile.kt
@@ -1,3 +1,0 @@
-package io.github.ludorival.pactjvm.mock.spring.providers.userservice
-
-data class UserProfile(val id: Long, val name: String, val email: String, val userPreferences: UserPreferences)
