--- pact-jvm-mock-spring/src/test/kotlin/io/github/ludorival/pactjvm/mock/spring/providers/userservice/UserServiceApplication.kt
@@ -1,14 +1,0 @@
-package io.github.ludorival.pactjvm.mock.spring.providers.userservice
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.context.annotation.ComponentScan
-@SpringBootApplication
-@ComponentScan(basePackages = ["io.github.ludorival.pactjvm.mock.spring.providers.userservice"])
-open class UserServiceApplication {
-    companion object {
-        @JvmStatic
-        fun main(args: Array<String>) {
-            org.springframework.boot.runApplication<UserServiceApplication>(*args)
-        }
-    }
-}
