--- pact-jvm-mock-spring/src/test/kotlin/io/github/ludorival/pactjvm/mock/spring/test/ShoppingListAppDeterministicTest.kt
@@ -1,8 +1,0 @@
-package io.github.ludorival.pactjvm.mock.spring.test
-
-import io.github.ludorival.pactjvm.mock.spring.DeterministicPact
-import io.github.ludorival.pactjvm.mock.PactConsumer
-import org.junit.jupiter.api.extension.ExtendWith
-
-@PactConsumer(DeterministicPact::class)
-open class ShoppingListAppDeterministicTest : ShoppingListAppTest()
