+++ pact-jvm-mock-test/src/main/kotlin/io/github/ludorival/pactjvm/mock/test/consumer/extension.kt
@@ -1,0 +1,7 @@
+package io.github.ludorival.pactjvm.mock.test.consumer
+
+import org.springframework.http.ResponseEntity
+
+ fun <T> ResponseEntity<T>.safeValue(lazyMessage: () -> String = { "Expect to have a non null value"}): T {
+    return body ?: error(lazyMessage())
+}
