+++ pact-jvm-mock-test/src/main/kotlin/io/github/ludorival/pactjvm/mock/test/userservice/UserPreferences.kt
@@ -1,0 +1,3 @@
+package io.github.ludorival.pactjvm.mock.test.userservice
+
+data class UserPreferences(val preferredShoppingListId: Long)
