+++ pact-jvm-mock-test/src/main/kotlin/io/github/ludorival/pactjvm/mock/test/userservice/UserProfile.kt
@@ -1,0 +1,3 @@
+package io.github.ludorival.pactjvm.mock.test.userservice
+
+data class UserProfile(val id: Long, val name: String, val email: String, val userPreferences: UserPreferences)
