+++ pact-jvm-mock-test/src/test/kotlin/io/github/ludorival/pactjvm/mock/test/ShoppingListAppDeterministicTest.kt
@@ -1,0 +1,7 @@
+package io.github.ludorival.pactjvm.mock.test
+
+import io.github.ludorival.pactjvm.mock.PactConsumer
+import org.junit.jupiter.api.extension.ExtendWith
+
+@PactConsumer(DeterministicPact::class)
+open class ShoppingListAppDeterministicTest : ShoppingListAppTest()
