+++ pact-jvm-mock-test/src/test/resources/pacts-deterministic/shopping-list-user-service.json
