+++ pact-jvm-mock-test/src/test/resources/pacts/shopping-list-shopping-service.json
