+++ pact-jvm-mock-test/src/test/resources/pacts/shopping-list-user-service.json
