+++ settings.gradle.kts
@@ -2,3 +2,4 @@ rootProject.name = "pact-jvm-mock"
 include("pact-jvm-mock")
 include("pact-jvm-mock-mockk")
 include("pact-jvm-mock-spring")
+include("pact-jvm-mock-test")
